+++ packages/router/src/grpc/view-protocol-server/broadcast-transaction.test.ts
@@ -11,7 +11,7 @@ import { Transaction } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/co
 import { broadcastTransaction } from './broadcast-transaction';
 import type { Services } from '@penumbra-zone/services';
 import { TransactionId } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/txhash/v1/txhash_pb';
-import { IndexedDbMock, MockServices, TendermintMock } from './test-utils';
+import { IndexedDbMock, MockServices, TendermintMock } from '../test-utils';
 
 const mockSha256 = vi.hoisted(() => vi.fn());
 vi.mock('@penumbra-zone/crypto-web', () => ({
