+++ packages/router/src/grpc/view-protocol-server/owned-position-ids.test.ts
@@ -9,7 +9,7 @@ import {
   OwnedPositionIdsRequest,
   OwnedPositionIdsResponse,
 } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/view/v1/view_pb';
-import { IndexedDbMock, MockServices } from './test-utils';
+import { IndexedDbMock, MockServices } from '../test-utils';
 import { Services } from '@penumbra-zone/services';
 import { PositionId } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/component/dex/v1/dex_pb';
 import { ownedPositionIds } from './owned-position-ids';
