+++ packages/router/src/grpc/view-protocol-server/unclaimed-swaps.test.ts
@@ -10,7 +10,7 @@ import {
   UnclaimedSwapsRequest,
   UnclaimedSwapsResponse,
 } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/view/v1/view_pb';
-import { IndexedDbMock, MockServices } from './test-utils';
+import { IndexedDbMock, MockServices } from '../test-utils';
 import { Services } from '@penumbra-zone/services';
 import { unclaimedSwaps } from './unclaimed-swaps';
