+++ packages/router/src/grpc/view-protocol-server/wallet-id.test.ts
@@ -7,7 +7,7 @@ import { createContextValues, createHandlerContext, HandlerContext } from '@conn
 import { ViewService } from '@buf/penumbra-zone_penumbra.connectrpc_es/penumbra/view/v1/view_connect';
 import { extLocalCtx } from '../../ctx';
 import { walletId } from './wallet-id';
-import { MockExtLocalCtx } from './test-utils';
+import { MockExtLocalCtx } from '../test-utils';
 
 describe('WalletId request handler', () => {
   let mockExtLocalCtx: MockExtLocalCtx;
